feat: reuse `wrap_sqlalchemy_exception`
