+++ advanced_alchemy/__init__.py
@@ -8,17 +8,21 @@ from advanced_alchemy.config import (
     SQLAlchemySyncConfig,
     SyncSessionConfig,
 )
+from advanced_alchemy.exceptions import (
+    IntegrityError,
+    MultipleResultsFoundError,
+    NotFoundError,
+    RepositoryError,
+    wrap_sqlalchemy_exception,
+)
+from advanced_alchemy.filters import FilterTypes
 from advanced_alchemy.repository._async import SQLAlchemyAsyncRepository
 from advanced_alchemy.repository._sync import SQLAlchemySyncRepository
-from advanced_alchemy.repository._util import wrap_sqlalchemy_exception
 from advanced_alchemy.repository.memory._async import SQLAlchemyAsyncMockRepository
 from advanced_alchemy.repository.memory._sync import SQLAlchemySyncMockRepository
 from advanced_alchemy.repository.typing import ModelT
 from advanced_alchemy.service._async import SQLAlchemyAsyncRepositoryReadService, SQLAlchemyAsyncRepositoryService
 from advanced_alchemy.service._sync import SQLAlchemySyncRepositoryReadService, SQLAlchemySyncRepositoryService
-
-from .exceptions import IntegrityError, MultipleResultsFoundError, NotFoundError, RepositoryError
-from .filters import FilterTypes
 
 __all__ = (
     "IntegrityError",
