+++ advanced_alchemy/exceptions.py
@@ -1,6 +1,10 @@
 from __future__ import annotations
 
+from contextlib import contextmanager
 from typing import Any
+
+from sqlalchemy.exc import IntegrityError as SQLAlchemyIntegrityError
+from sqlalchemy.exc import MultipleResultsFound, SQLAlchemyError
 
 from advanced_alchemy.utils.deprecation import deprecated
 
@@ -88,3 +92,30 @@ class NotFoundError(RepositoryError):
 
 class MultipleResultsFoundError(AdvancedAlchemyError):
     """A single database result was required but more than one were found."""
+
+
+@contextmanager
+def wrap_sqlalchemy_exception() -> Any:
+    """Do something within context to raise a `RepositoryError` chained
+    from an original `SQLAlchemyError`.
+
+        >>> try:
+        ...     with wrap_sqlalchemy_exception():
+        ...         raise SQLAlchemyError("Original Exception")
+        ... except RepositoryError as exc:
+        ...     print(f"caught repository exception from {type(exc.__context__)}")
+        ...
+        caught repository exception from <class 'sqlalchemy.exc.SQLAlchemyError'>
+    """
+    try:
+        yield
+    except MultipleResultsFound as e:
+        msg = "Multiple rows matched the specified key"
+        raise MultipleResultsFoundError(msg) from e
+    except SQLAlchemyIntegrityError as exc:
+        raise IntegrityError from exc
+    except SQLAlchemyError as exc:
+        msg = f"An exception occurred: {exc}"
+        raise RepositoryError(msg) from exc
+    except AttributeError as exc:
+        raise RepositoryError from exc
